+++ src/main/java/components/control/TextControl.java
@@ -14,8 +14,8 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 public final class TextControl extends Control {
     private static final Logger LOGGER = LogManager.getLogger();
     private int timeOut = 60;
-    private int waitAfterClean = 0;
-    private int waitForClick = 0;
+    private long waitAfterClean = 0;
+    private long waitForClick = 0;
     private boolean cleanText = true;
 
     /**
